[#129687459] can also remove archive when removing mod
Also offering to disable dependent mods but this isn't implemented yet.

+++ src/extensions/mod_management/views/ModRow.tsx
@@ -1,18 +1,26 @@
+import { showDialog } from '../../../actions/notifications';
+import { IComponentContext } from '../../../types/IComponentContext';
+import { DialogActions, DialogType, IDialogContent, IDialogResult } from '../../../types/IDialog';
 import { IIconDefinition } from '../../../types/IIconDefinition';
 import getAttr from '../../../util/getAttr';
 import IconBar from '../../../views/IconBar';
 
 import { connect } from '../../../util/ComponentEx';
 
+import { removeDownload } from '../../download_management/actions/state';
 import { IProfileMod } from '../../profile_management/types/IProfile';
 
 import { removeMod } from '../actions/mods';
 import { IMod } from '../types/IMod';
 import { IModAttribute } from '../types/IModAttribute';
 
+import * as Promise from 'bluebird';
 import * as fs from 'fs-extra-promise';
 import * as React from 'react';
 import { Checkbox } from 'react-bootstrap';
+
+import {log} from '../../../util/log';
+import * as util from 'util';
 
 export interface IBaseProps {
   mod: IMod;
@@ -26,11 +34,18 @@ export interface IBaseProps {
 
 interface IActionProps {
   onRemoveMod: (modId: string) => void;
+  onShowDialog: (type: DialogType, title: string, content: IDialogContent,
+                 actions: DialogActions) => Promise<IDialogResult>;
 }
 
 type IProps = IBaseProps & IActionProps;
 
 class ModRow extends React.Component<IProps, {}> {
+  public static contextTypes: React.ValidationMap<any> = {
+    api: React.PropTypes.object.isRequired,
+  };
+
+  public context: IComponentContext;
 
   private modActions: IIconDefinition[];
 
@@ -81,11 +96,38 @@ class ModRow extends React.Component<IProps, {}> {
   }
 
   private remove = () => {
-    const { mod, onRemoveMod } = this.props;
+    const { mod, onRemoveMod, onShowDialog } = this.props;
     if (mod.installationPath) {
-      fs.removeAsync(mod.installationPath)
+      let removeMod = true;
+      let removeArchive = false;
+      let disableDependent = false;
+      onShowDialog('question', 'Confirm Removal', {
+        checkboxes: [
+          { id: 'mod', text: 'Remove Mod', value: true },
+          { id: 'archive', text: 'Remove Archive', value: false },
+          { id: 'dependents', text: 'Disable Dependent', value: false },
+        ],
+      }, {
+          Cancel: null,
+          Continue: null,
+        }).then((result: IDialogResult) => {
+          log('info', 'res', util.inspect(result));
+          removeMod = result.action === 'Continue' && result.input.mod;
+          removeArchive = result.action === 'Continue' && result.input.archive;
+          disableDependent = result.action === 'Continue' && result.input.dependents;
+          if (removeMod) {
+            return fs.removeAsync(mod.installationPath);
+          } else {
+            return Promise.resolve();
+          }
+        })
         .then(() => {
-          onRemoveMod(mod.id);
+          if (removeMod) {
+            onRemoveMod(mod.id);
+          }
+          if (removeArchive) {
+            this.context.api.events.emit('remove-download', mod.archiveId);
+          }
         });
     } else {
       onRemoveMod(mod.id);
@@ -127,8 +169,9 @@ function mapStateToProps(state) {
 
 function mapDispatchToProps(dispatch: Redux.Dispatch<any>): IActionProps {
   return {
-    onRemoveMod: (modId: string) =>
-      dispatch(removeMod(modId)),
+    onRemoveMod: (modId: string) => dispatch(removeMod(modId)),
+    onShowDialog:
+      (type, title, content, actions) => dispatch(showDialog(type, title, content, actions)),
   };
 }
